requirements.txt
appdirs==1.4.0
Django==1.10.5
django-mptt==0.8.7
django-river==0.9.0
factory-boy==2.8.1
Faker==0.7.7
mock==2.0.0
packaging==16.8
pbr==1.10.0
psycopg2==2.6.2
python-dateutil==2.6.0
six==1.10.0
